Week 7 assignment. Classes for State, Player, and Die were created for the assigment, and an exception for rolling a one.
I couldn't add the results of the Die() value to the turn value, so I got lost trying to do the math adding the dice returned values to the turn totals and the game totals. Also, I could't break out of while and IF statements to end the game or proceed futher.